requirements.txt
asgiref==3.2.5
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bleach==3.1.4
click==7.1.1
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-background-tasks==1.2.5
django-class==2019.4.13
django-compat==1.0.15
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-isbn-field==0.5.2
django-phonenumber-field==4.0.0
entrypoints==0.3
gunicorn==20.0.4
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.5
nltk==3.4.5
notebook==6.0.3
numpy==1.18.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
phonenumbers==8.11.5
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psycopg2==2.8.4
ptyprocess==0.6.0
public==2019.4.13
Pygments==2.6.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-stdnum==1.13
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
scikit-surprise==1.1.0
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
surprise==0.1
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
zipp==3.1.0
